# api/openapi.yaml
openapi: 3.1.0
info:
  title: MiniGBA Cloud Save API
  version: 1.0.0
  description: Versioned cloud saves for the MiniGBA WeChat Mini Program. ROM content is never accepted.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
paths:
...
  /v1/auth/wechat/login:
    post:
      security: []
      operationId: wechatLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, clientDeviceId, deviceName]
              properties:
                code: { type: string, minLength: 1 }
                clientDeviceId: { type: string, format: uuid }
                deviceName: { type: string, minLength: 1, maxLength: 80 }
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/logout:
    post:
      operationId: logout
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/refresh:
    post:
      operationId: refreshAccessToken
      responses:
        '200':
          description: Access token renewed for the current session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Error' }
...
  /v1/account/deletion:
    post:
      operationId: requestAccountDeletion
      parameters:
        - name: X-Confirm-Account-Deletion
          in: header
          required: true
          schema: { type: string, const: DELETE }
      responses:
        '202':
          description: Deletion queued and all sessions revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionStatus' }
    get:
      operationId: getAccountDeletionStatus
      description: The original access token acts only as a short-lived deletion receipt after sessions are revoked.
      responses:
        '200':
          description: Latest deletion job status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionStatus' }
        '404': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresAt, deviceId]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, const: Bearer }
        expiresAt: { type: string, format: date-time }
        deviceId: { type: string, format: uuid }
    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresAt]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, const: Bearer }
        expiresAt: { type: string, format: date-time }
...
    DeletionStatus:
      type: object
      required: [jobId, status, attempts, createdAt, updatedAt]
      properties:
        jobId: { type: string, format: uuid }
        status: { type: string, enum: [pending, running, complete, failed] }
        attempts: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
    Error:
      description: Structured error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }